Added install of Sun and Abraham code and fixed S&A event study graph
The fix to the S&A graph involved setting endpoint values to be equal to the max and min points that we are looking at in the graph. Otherwise, they get grouped in with the omitted (treatment time) observations. I included a link to a Goodman-Bacon tweet about the issue.